requirements.txt
Flask==3.1.0
requests==2.26.0
sseclient-py==1.7.2
python-dotenv==0.19.0
pydantic==1.8.2
asyncio==3.4.3
aiohttp==3.7.4.post0
websockets==9.1
psycopg2-binary==2.9.1
numpy